refactor(components): extract FormContainer and Button
Extracted reusable FormContainer and Button components, removing styles from Login's styles.js.
