+++ src/components/Button/index.jsx
@@ -1,0 +1,10 @@
+import { Children } from "react";
+import Container from "./styles";
+
+export const Button = ({ type, children, disable }) => {
+  return (
+    <Container type={type} disabled={disable}>
+      {children}
+    </Container>
+  );
+};
