+++ src/components/FormContainer/index.jsx
@@ -1,0 +1,5 @@
+import Container from "./styles";
+
+export const FormContainer = ({ children }) => {
+  return <Container>{children}</Container>;
+};
