+++ src/views/Login/styles.js
@@ -75,37 +75,3 @@ export default styled.div`
     }
   }
 `;
-
-export const Form = styled.form`
-  display: flex;
-  flex-direction: column;
-  row-gap: 17px;
-  margin: 0 auto;
-  @media (min-width: 768px) {
-    width: 400px;
-    row-gap: 32px;
-    .containerButton {
-      max-width: 400px;
-    }
-  }
-`;
-
-export const Button = styled.button`
-  background: radial-gradient(
-    circle,
-    ${colors.gradient.lighterBlue} 50%,
-    ${colors.gradient.darkerBlueButton} 100%
-  );
-  padding: 8px 24px;
-  border: none;
-  border-radius: 5px;
-  color: white;
-  font: 500 20px "Raleway", sans-serif;
-  float: right;
-  cursor: pointer;
-  @media (min-width: 768px) {
-    width: 100%;
-    float: none;
-    padding: 12px 24px;
-  }
-`;
